refactor: package rooms together
