+++ HospitalSystem.ConsoleUI/Secretary/EquipUI.cs
@@ -1,4 +1,5 @@
 using HospitalSystem.Core;
+using HospitalSystem.Core.Rooms;
 using HospitalSystem.ConsoleUI.Director;
 
 namespace HospitalSystem.ConsoleUI;
