+++ HospitalSystem.Core/Equipment/EquipmentBatchService.cs
@@ -1,3 +1,5 @@
+using HospitalSystem.Core.Rooms;
+
 namespace HospitalSystem.Core;
 
 [System.Serializable]
