+++ HospitalSystem.Core/Rooms/RoomRepository.cs
@@ -1,7 +1,7 @@
 using System.Linq.Expressions;
 using MongoDB.Driver;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Rooms;
 
 public class RoomRepository : IRoomRepository
 {
